On disk cache and package reorg
